Enable the developer to inject values for the new  geometry created with CutPolygonCADTool.
This is a provisory solution. We use the same listener to warn about all changes, which can be differentiated by means of the key used in the listener. Likely a better solution will be having several listener. How that could be implemented should be discussed (several listeners shared for all cadtools, especific listeners for every cadtool, etc).

Example (not to use in production code, please, first check types and so):

public class YourClass implements EndGeometryListener {

    public void endGeometry(FLayer layer, String cadToolKey){
        CADTool cadTool = CADExtension.getCADTool();
        Value[] values = getYourValuesAsYouWant();
        ((CutPolygonCADTool) cadTool).setParametrizableValues(values);
    }

}