Improve warning dialogs of ExportTo
We have observed that the export window, where the number of features
is shown, appears only when there are some features selected. If
nothing is selected, then the "file save" dialog is shown directly.

In order to show the name of the layer, even if nothing is selected,
the export layer window is always shown, with the name of the
layer (buildings.shp in the following example) and the number of
selected and total features:

    buildings.shp
    Number of features selected for export: 1/13573

And, if nothing is selected:

    buildings.shp
    Number of features selected for export: 13573/13573