Uses a simpler method to generate a GeneralPathX from a IGeometry
Use the constructor public GeneralPathX(Shape s) instead of recreate
the path is a simpler. The only significant difference between both
is that the previous method iterate the points flattening it, and
the new one, not flattens the arcs.

But, as the arc tools in gvSIG are already flatening the paths in the
endGeometry method doing it again, is in general pointless